Add documentation for gentest (and a small bugfix) (#21)
* Add documentation for gentest package

Previously, there wasn't any documentation for the gentest package.
We want new users to Genny to be able to test their Generators easily,
so adding documentation will help them write tests more easily.

* Fix printing of CompareCommands failure output

When CompareCommands failed, its error message had the the expectation
and the actual values reversed. This led to some minor confusion. This
reverses the arguments to be correct so as not to confuse people.